core/views/CustomUserView.py
# https://studygyaan.com/django/django-rest-framework-tutorial-register-login-logout
from django.contrib.auth import authenticate
from django.core.exceptions import ObjectDoesNotExist
from core.models import CustomUser
from core.serializers import CustomUserSerializer
from rest_framework import status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.response import Response
from rest_framework.authtoken.models import Token
from rest_framework.permissions import IsAuthenticated


@api_view(['POST'])
def register_user(request):
    if request.method == 'POST':
        serializer = CustomUserSerializer(data=request.data)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data, status=status.HTTP_201_CREATED)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)