requirements.txt
asgiref==3.7.2
coverage==7.4.4
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.3.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.1.1
pytest-cov==4.1.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
